search: move SearchTag to components/search_tag
